Move FileTrait.php into Traits directory
